refactor: extract GmlPortrayalCatalogueBase from 8 identical catalogues
Introduce GmlPortrayalCatalogueBase in Portrayals that provides all shared
infrastructure: palette loading, rule discovery, XSLT compilation with
caching, symbol/line-style/area-fill loading, and configurable XmlResolver.

Each per-spec catalogue is now a thin subclass:
- S-122, S-124, S-127, S-128: only ProductSpec override
- S-125, S-421: add FetchRuleFallbackXmlResolver
- S-411: adapter rule injection, simplified palette loading, single-rule filter
- S-129: segment-based path resolver for sub-directory XSLT includes
